refactoring and moving component to qmongjs
