+++ packages/qmongjs/.lintstagedrc
@@ -1,8 +1,4 @@
 {
-  "*.+(js|ts|tsx)": [
-    "eslint"
-  ],
-  "**/*.+(js|jsx|ts|tsx|json|html|yml|yaml|css|md)": [
-    "prettier --write"
-  ]
+  "*.+(js|ts|tsx)": ["eslint"],
+  "**/*.+(js|jsx|ts|tsx|json|html|yml|yaml|css|md)": ["prettier --write"]
 }
